Menambahkan get user discussion

+++ src/controller/discussionController.js
@@ -2,16 +2,20 @@ const { ObjectId } = require('mongodb');
 const Discussion = require('../models/discussion.js');
 const Category = require('../models/discussionCategory.js');
 const DiscussionReply = require('../models/discussionReply.js');
+const User = require('../models/user.js');
 
 const createDiscussion = async(req,res) =>{
+    const user = req.user;
     const discussionID = new ObjectId();
     const {
-        name,
         title,
         discussion,
         categories
     } = req.body;
 
+    const nameuser = await User.findOne({
+        _id: user
+    })
     const newDate = new Date();
     const monthID = ['Januari', 'Februari', 'Maret', 'April', 'Mei', 'Juni', 'Juli', 'Agustus', 'September', 'Oktober', 'November', 'Desember'];
     const date = `${newDate.getDate()} ${monthID[newDate.getMonth()]} ${newDate.getFullYear()}`;
@@ -41,7 +45,8 @@ const createDiscussion = async(req,res) =>{
 
     const newDiscussion = new Discussion({
         _id: discussionID,
-        name,
+        userid: user,
+        username: nameuser.username,
         date,
         title,
         categories,
@@ -79,8 +84,22 @@ const getDiscussion = async(req, res) => {
     })
 }
 
+const getUserDiscussion = async(req, res) => {
+    const user = req.user;
+    
+    const discussion = await Discussion.find({
+        userid: user._id
+    })
+
+    res.json({
+        success: true,
+        data: discussion
+    })
+}
+
 const editDiscussion = async(req, res, next) => {
     const { id } = req.params;
+    const user = req.user;
     const {
         title,
         categories,
@@ -88,6 +107,17 @@ const editDiscussion = async(req, res, next) => {
         isSolved
     } = req.body;
 
+    const discussionId = await Discussion.findOne({
+        _id: id
+    })
+
+    if(user._id !== discussionId.userid) {
+        res.status(400).json({
+            success: false,
+            message: 'Tidak dapat melakukan edit pada diskusi ini'
+        })
+        return
+    }
     const newDate = new Date();
     const monthID = ['Januari', 'Februari', 'Maret', 'April', 'Mei', 'Juni', 'Juli', 'Agustus', 'September', 'Oktober', 'November', 'Desember'];
     const date = `${newDate.getDate()} ${monthID[newDate.getMonth()]} ${newDate.getFullYear()}`;
@@ -144,16 +174,18 @@ const editDiscussion = async(req, res, next) => {
         count++;
     }
 
-    const discuss = await Discussion.findOneAndUpdate({
-        _id: id,
-        $set: {
-            title,
-            categories,
-            discussion,
-            date,
-            isSolved
+    const discuss = await Discussion.findOneAndUpdate(
+        {_id: id},
+        {
+            $set: {
+                title,
+                categories,
+                discussion,
+                date,
+                isSolved
+            }
         }
-    });
+    );
 
     if(discuss) {
         res.status(201).json({
@@ -165,7 +197,20 @@ const editDiscussion = async(req, res, next) => {
 }
 
 const deleteDiscussion = async(req, res) => {
-    const {id} = req.params;
+    const { id } = req.params;
+    const user = req.user;
+
+    const discussionId = await Discussion.findOne({
+        _id: id
+    })
+
+    if(user._id !== discussionId.userid) {
+        res.status(400).json({
+            success: false,
+            message: 'Tidak dapat menghapus diskusi ini'
+        })
+        return
+    }
 
     const discussion = await Discussion.findOneAndDelete({
         _id: id
@@ -201,6 +246,7 @@ const deleteDiscussion = async(req, res) => {
 module.exports = {
     createDiscussion,
     getAllDiscussion,
+    getUserDiscussion,
     getDiscussion,
     editDiscussion,
     deleteDiscussion
